% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dashboardPlot}
\alias{dashboardPlot}
\title{Create interactive Shiny app to explore results}
\usage{
dashboardPlot(all.glm.res)
}
\arguments{
\item{all.glm.res}{Provide the object created in the function excessCases}
}
\value{
Launches an interactive shiny app
}
\description{
\code{dashboardPlot} Creates an interactive Shiny plot
to explore results generated in the function excessCases. 
Drop down menus allow for viewing different syndromes, age groups,
and geographies, and for looking at plots of raw counts,
proportions, or Observed/Expected
}
