% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPlotOe.R
\name{dashboardPlotOe}
\alias{dashboardPlotOe}
\title{Create interactive Shiny app to view Observed/Expected output over time by geography and by age}
\usage{
dashboardPlotOe(excess_output, datevar, agevar, statevar, outcome, yaxis,
  facet)
}
\arguments{
\item{excess_output}{Provide the object returned by the function
\code{\link{excessCases}}}

\item{datevar}{Provide the datevar used in the function
\code{\link{excessCases}}}

\item{agevar}{Provide the agevar used in the function
\code{\link{excessCases}}}

\item{statevar}{Provide the statevar geographic variable used in the function
\code{\link{excessCases}}}

\item{outcome}{Provide one of the outcomes / syndromes of interest
\code{\link{excessCases}}}

\item{yaxis}{Provide the variable intended for the y-axis (either agevar or statevar)
\code{\link{excessCases}}}

\item{facet}{Provide the variable intended for the facet groupings (either agevar or statevar)
\code{\link{excessCases}}}
}
\value{
Returns an object representing the shinyapp. Depending on the
  environment (RStudio vs. console) the app may then be passed to
  \code{print()}, which will start the server. Refer to
  \code{link[shiny]{shinyApp}} for more details.
}
\description{
\code{dashboardPlotOe} Creates an interactive Shiny plot to explore Observed/Expected 
  results generated in the function \code{\link{excessCases}}. Drop down menus
  allow for viewing Observed/Expected by different age groups and geographies.
}
