% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excessExtract}
\alias{excessExtract}
\title{ONE LINER DESCRIPTION}
\usage{
excessExtract(ds, syndrome, extract.quantity)
}
\arguments{
\item{ds}{NEEDS DOCUMENTATION}

\item{syndrome}{NEEDS DOCUMENTATION}

\item{extract.quantity}{NEEDS DOCUMENTATION}
}
\value{
NEEDS DOCUMENTATION.
}
\description{
\code{excessExtract} DOES SOMETHING
}
\details{
EXTENDED DESCRIPTION, THE FOLLOWING IS AN EXAMPLE: This is a generic
function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the
arguments \code{...} should be unnamed, and dispatch is on the first
argument.
}
