% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{mostRecentTimestamp}
\alias{mostRecentTimestamp}
\title{Identify the timestamp of the most recently modified file in the file cache}
\usage{
mostRecentTimestamp(fname, basepath = ".")
}
\arguments{
\item{fname}{The name of the directory in \code{basepath} where various
revisions of the file are stored. I.e., \code{file.txt} should be a
directory, with revisions of the true \code{file.txt} stored inside of
it.}

\item{basepath}{A string. The path which stores the \code{fname} directory.
Default '.'}
}
\value{
A POSIXct object specifying the \code{mtime} of the most recently
  modified file in \code{basepath/fname}
}
\description{
Returns the timestamp of the most recently modified file. If no such file
exists, or if the directory \code{basepath/fname} doesn't exist, returns NA.
}
\examples{
library(lubridate)

saveRDS(mtcars, 'cars')
saveRDS(mtcars, 'cars')

# Some time elapses...

# Decide if the latest version of 'mtcars' is "too old"
if (mostRecentTimestamp('mtcars') \%--\% now() > hours(24)) {
  # Store a "new" version
  saveRDS(mtcars, 'cars')
} else {
  cached_mtcars <- retrieveRDS('mtcars')
}

}
