% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{retrievePath}
\alias{retrievePath}
\title{Retrieve the path of a file modified nearest to a specific date}
\usage{
retrievePath(fname, basepath = ".", goalDate = lubridate::now())
}
\arguments{
\item{fname}{The name of the directory in \code{'basepath'} where various
revisions of the file are stored. I.e., \code{'file.txt'} should be a
directory, with revisions of the true \code{'file.txt'} stored inside of
it.}

\item{basepath}{A string. The path which stores 'basepath'. Default '.'}

\item{goalDate}{The 'goal' date for the revision we are trying to retrieve.
Whichever file is closest in time to the 'goal' date will be selected}
}
\value{
A string. The full path to the file, or an error if such a file
  cannot be found.
}
\description{
Retrieves the path of a file stored in \code{'basepath'/'fname'} that has
the closest modification date to 'goalDate'. If no such file exists, an
error will be thrown.
}
\examples{

storeRDS(mtcars, 'mtcars')
retrievePath('mtcars', goalDate=lubridate::now() - lubridate::days(10))
}
