% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{retrieveRDS}
\alias{retrieveRDS}
\title{Retrieve an RDS modified nearest to a specific date}
\usage{
retrieveRDS(fname, basepath = ".", goalDate = Sys.time())
}
\arguments{
\item{fname}{The name of the directory in \code{basepath} where various
revisions of the file are stored. I.e., \code{file.txt} should be a
directory, with revisions of the true \code{file.txt} stored inside of
it.}

\item{basepath}{A string. The path which stores \code{fname}. Default
\code{'.'}}

\item{goalDate}{The 'goal' date for the revision we are trying to retrieve.
Whichever file is closest in time to the 'goal' date will be selected}
}
\description{
Retrieves an RDS file stored in \code{basepath/fname} that has
the closest modification date to \code{goalDate}. If no such file exists, an
error will be thrown.
}
\examples{
storeRDS(mtcars, 'mtcars')
identical(mtcars, retrieveRDS('mtcars'))

storeRDS(mtcars, 'mtcars', basepath='.')
mtcars2 <- mtcars
mtcars2[,'cyl'] <- 1
storeRDS(mtcars2, 'mtcars')
mtcars_retrieved <- retrieveRDS('mtcars', goalDate = lubridate::now())
identical(mtcars, mtcars_retrieved)

}
