% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ts_format}
\alias{ts_format}
\title{Format line list data into time series}
\usage{
ts_format(
  line.list,
  datevar,
  statevar,
  sub.statevar,
  agevar,
  syndromes,
  resolution = "day",
  remove.final = F
)
}
\arguments{
\item{line.list}{Input data to be reformatted. NEEDS DETAILED DESCRIPTION}

\item{datevar}{A string. What variable contains the date?}

\item{statevar}{A string. What variable contains the state?}

\item{sub.statevar}{A string. NEEDS DOCUMENTATION}

\item{agevar}{A string. What variable contains the age group? Use 'none'
if there is no age grouping in the data}

\item{syndromes}{A character vector. Which variables contain counts of
syndromic data?}

\item{resolution}{One of \code{c("day", "week", "month")}. What is the data
binned by?}

\item{remove.final}{A logical scalar. NEEDS DOCUMENTATION.}
}
\value{
NEEDS DOCUMENTATION.
}
\description{
\code{ts_format} DOES SOMETHING
}
\details{
EXTENDED DESCRIPTION, THE FOLLOWING IS AN EXAMPLE: This is a generic
function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the
arguments \code{...} should be unnamed, and dispatch is on the first
argument.
}
\examples{
 n.obs <- 10000
 set.seed(42)

 simulated_data <-
   as.data.frame(matrix(NA, nrow=n.obs, ncol=5))

 names(sim1) <- c('state','date','agegrp','ili','resp')

 sim1$state<- c( rep('CT', times=n.obs*0.3),
                 rep("NY", times=n.obs*0.7) )

 sim1$agegrp <- sample(1:5, n.obs, replace=T)
 sim1$date   <- sample(seq.Date(from=as.Date('2019-01-01'),
                                by='day',
                                length.out=500), 
                       1000,
                       replace=T)

 sim1$ili  <- rbinom(n=n.obs, size=1, prob=0.05)
 sim1$resp <- rbinom(n=n.obs, size=1, prob=0.1)

 ts1 <- ts_format(line.list=sim1,
                  datevar='date',
                  agevar='agegrp',
                  statevar='state',
                  syndromes=c('ili','resp'))

}
