% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPlot.R
\name{dashboardPlot}
\alias{dashboardPlot}
\title{Create interactive Shiny app to explore results}
\usage{
dashboardPlot(all.glm.res)
}
\arguments{
\item{all.glm.res}{Provide the object returned by the function
\code{\link{excessCases}}}
}
\value{
Returns an object representing the shinyapp. Depending on the
  environment (RStudio vs. console) the app may then be passed to
  \code{print()}, which will start the server. Refer to
  \code{link[shiny]{shinyApp}} for more details.
}
\description{
\code{dashboardPlot} Creates an interactive Shiny plot to explore results
  generated in the function \code{\link{excessCases}}. Drop down menus
  allow for viewing different syndromes, age groups, and geographies, and
  for looking at plots of raw counts, proportions, or Observed/Expected.
}
