% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excessCases}
\alias{excessCases}
\title{ONE LINER DESCRIPTION}
\usage{
excessCases(
  ds,
  sub.statevar = "none",
  statevar = "state",
  agevar = "none",
  datevar,
  use.syndromes,
  denom.var,
  flu.import = T,
  rsv.import = T,
  adj.flu = F,
  adj.rsv = F,
  flu.var = "flu.var",
  rsv.var = "rsv.var",
  time.res = "day",
  extrapolation.date
)
}
\arguments{
\item{sub.statevar}{A string. NEEDS DOCUMENTATION}

\item{statevar}{A string. What variable contains the state?}

\item{agevar}{A string. What variable contains the age group? Use 'none'
if there is no age grouping in the data}

\item{datevar}{A string. What variable contains the date?}

\item{use.syndromes}{NEEDS DOCUMENTATION.}

\item{denom.var}{NEEDS DOCUMENTATION.}

\item{flu.import}{A logical scalar. NEEDS DOCUMENTATION.}

\item{rsv.import}{A logical scalar. NEEDS DOCUMENTATION.}

\item{adj.flu}{A logical scalar. NEEDS DOCUMENTATION.}

\item{adj.rsv}{A logical scalar. NEEDS DOCUMENTATION.}

\item{flu.var}{A string. NEEDS DOCUMENTATION.}

\item{rsv.var}{A string. NEEDS DOCUMENTATION.}

\item{time.res}{One of \code{c("day", "week", "month")}. What is the data
binned by?}

\item{extrapolation.date}{NEEDS DOCUMENTATION.}
}
\value{
NEEDS DOCUMENTATION.
}
\description{
\code{excessCases} DOES SOMETHING
}
\details{
EXTENDED DESCRIPTION, THE FOLLOWING IS AN EXAMPLE: This is a generic
function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the
arguments \code{...} should be unnamed, and dispatch is on the first
argument.
}
\examples{
 ili.data <- ilinet(region = c("state"))
 ili.data$state <- state.abb[match(ili.data$region, state.name)]

 ili.data <- ili.data[, c("state", "week_start", "ilitotal", "total_patients")]
 ili.data <- ili.data[ili.data$state \%in\% c("CA", "NY", "WA", "NJ", "CT"), ]

 excess_cases <-
   excessCases(ds = ili.data,
               datevar = "week_start",
               agevar = "none",
               statevar = "state",
               denom.var = "total_patients",
               use.syndromes = c("ilitotal"),
               time.res = "week")
}
