% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{excessExtract}
\alias{excessExtract}
\title{ONE LINER DESCRIPTION}
\usage{
excessExtract(ds, syndrome, extract.quantity)
}
\arguments{
\item{ds}{An object created by the function excessCases}

\item{syndrome}{Character. For which syndrome should results be extracted? The name for this variable should match a syndrom provided to excessCases}

\item{extract.quantity}{Which element of the output (e.g., 'lpi') from the output of excessCases do you want to extract. See the values in the help for excessCases to see the options.}
}
\value{
a multidimensional array with dimensions for time, age group, and geography (state, and sub-state)
}
\description{
\code{excessExtract} Helper function to format the results
from excessCases
}
\details{
EXTENDED DESCRIPTION This function extracts specific estimates
produced in excessCases and saves them in a multidimensional
array.
}
