% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCache.R
\name{mostRecentTimestamp}
\alias{mostRecentTimestamp}
\title{Identify the timestamp of the most recently modified file in the file cache}
\usage{
mostRecentTimestamp(fname, basepath = ".")
}
\arguments{
\item{fname}{The name of the directory in \code{basepath} where various
revisions of the file are stored. I.e., \code{file.txt} should be a
directory, with revisions of the true \code{file.txt} stored inside of
it.}

\item{basepath}{A string. The path which stores the \code{fname} directory.
Default '.'}
}
\value{
A POSIXct object specifying the \code{mtime} of the most recently
  modified file in \code{basepath/fname}
}
\description{
Given an R object, attempts to store it in the directory
\code{basepath/fname}. The name given to the file will be of the form
\code{DATE.rds}, where \code{DATE} is of the format %Y_%m_%d_%H_%M.
An error will be thrown if \code{basepath} does not exist. However,
if \code{basepath/fname} does not exist, an attempt will be made to create
it. The \code{DATE} is the current time.
}
\examples{
library(lubridate)

saveRDS(mtcars, 'cars')
saveRDS(mtcars, 'cars')

# Some time elapses...

# Decide if the latest version of 'mtcars' is "too old"
if (mostRecentTimestamp('mtcars') \%--\% now() > hours(24)) {
  # Store a "new" version
  saveRDS(mtcars, 'cars')
} else {
  cached_mtcars <- retrieveRDS('mtcars')
}

}
